+++ rts/RtsStartup.c
@@ -25,7 +25,6 @@
 #include "StgRun.h"
 #include "Prelude.h"            /* fixupRTStoPreludeRefs */
 #include "Adjustor.h"           /* initAdjustors */
-#include "ThreadLabels.h"
 #include "sm/BlockAlloc.h"
 #include "Trace.h"
 #include "StableName.h"
@@ -384,9 +383,6 @@ hs_init_ghc(int *argc, char **argv[], RtsConfig rts_config)
     /* initialise file locking, if necessary */
     initFileLocking();
 
-    /* initialise thread label table (tso->char*) */
-    initThreadLabelTable();
-
 #if defined(PROFILING)
     initProfiling();
 #endif
@@ -557,9 +553,6 @@ hs_exit_(bool wait_foreign)
 
     /* free the stable name table */
     exitStableNameTable();
-
-    /* free the thread label table */
-    freeThreadLabelTable();
 
 #if defined(PROFILING)
     reportCCSProfiling();