+++ rts/sm/NonMovingMark.c
@@ -990,6 +990,9 @@ trace_tso (MarkQueue *queue, StgTSO *tso)
     trace_trec_header(queue, tso->trec);
     markQueuePushClosure_(queue, (StgClosure *) tso->stackobj);
     markQueuePushClosure_(queue, (StgClosure *) tso->_link);
+    if (tso->label != NULL) {
+        markQueuePushClosure_(queue, (StgClosure *) tso->label);
+    }
     if (   tso->why_blocked == BlockedOnMVar
         || tso->why_blocked == BlockedOnMVarRead
         || tso->why_blocked == BlockedOnBlackHole